app/requirements.txt
Flask
mysql-connector
flask-mysqldb
mysqlclient
passlib
Flask-Script
flask-wtf
wtforms-html5